fix(avatar upload): delete only if old file exists

+++ src/routes/users/user/avatar/upload.ts
@@ -116,7 +116,7 @@ export default function (
                     await compress(file.path, user.id);
                     const oldFile = (
                         await bucket.find({ "metadata.id": user.id }).toArray()
-                    )[0];
+                    )?.[0];
                     await new Promise((resolve, reject) => {
                         const uploadStream = fs
                             .createReadStream(`tmp/avatars/${user.id}.png`)
@@ -128,7 +128,9 @@ export default function (
                         uploadStream.on("error", reject);
                         uploadStream.on("close", resolve);
                     });
-                    await bucket.delete(oldFile._id);
+                    if (oldFile) {
+                        await bucket.delete(oldFile._id);
+                    }
                 } catch (err) {
                     fastify.log.error(err);
                     res.code(422).send({
